REFACTOR: move Alarm class to alarm module
* Alarm class can now be imported from outside the package
