+++ ros_alarms/__init__.py
@@ -1,5 +1,5 @@
 from alarms import AlarmListener, AlarmBroadcaster, HeartbeatMonitor
-from alarms import parse_json_str
+from alarms import parse_json_str, Alarm
 
 try:
     from tx_alarms import TxAlarmListener, TxAlarmBroadcaster, TxHeartbeatMonitor
